Remove other development
