+++ src/main/webapp/app/shared/metis/post/post.component.html
@@ -48,4 +48,4 @@
         ></jhi-posting-content>
     </div>
 </div>
-<jhi-post-footer [modalRef]="modalRef" [previewMode]="previewMode" [posting]="posting" [showAnswers]="showAnswers"></jhi-post-footer>
+<jhi-post-footer [modalRef]="modalRef" [previewMode]="previewMode" [posting]="posting"></jhi-post-footer>
