+++ src/main/webapp/app/shared/metis/posting-footer/post-footer/post-footer.component.html
@@ -1,31 +1,14 @@
 <div class="me-0 mt-2 row justify-content-between" [ngClass]="previewMode ? 'mb-2' : ''">
+    <div class="col-auto row">
+        <!-- Post tags -->
+        <div class="col-auto">
+            <span *ngFor="let tag of tags" class="post-tag px-1 me-1">
+                {{ tag }}
+            </span>
+        </div>
+    </div>
     <!-- Post reactions -->
     <div *ngIf="!previewMode" class="col-auto ps-4 pe-0">
-        <jhi-post-reactions-bar
-            [posting]="posting"
-            [(showAnswers)]="showAnswers"
-            [sortedAnswerPosts]="sortedAnswerPosts"
-            (openPostingCreateEditModal)="createAnswerPostModal.open()"
-        ></jhi-post-reactions-bar>
-    </div>
-    <!-- Post tags -->
-    <div class="col-auto ps-4 pe-0">
-        <span *ngFor="let tag of tags" class="post-tag px-1 me-1">
-            {{ tag }}
-        </span>
+        <jhi-post-reactions-bar [posting]="posting"></jhi-post-reactions-bar>
     </div>
 </div>
-
-<ng-container *ngIf="showAnswers">
-    <jhi-answer-post *ngFor="let answerPost of sortedAnswerPosts" [posting]="answerPost"></jhi-answer-post>
-</ng-container>
-<!-- add new answer to expanded discussion -->
-<button *ngIf="showAnswers && sortedAnswerPosts.length > 0" class="btn btn-sm btn-outline-secondary mt-2" (click)="createAnswerPostModal.open()">
-    <fa-icon [icon]="faComments"></fa-icon>
-    <span class="d-none d-md-inline">{{ 'artemisApp.metis.answerNow' | artemisTranslate }}</span>
-</button>
-<jhi-answer-post-create-edit-modal
-    #createAnswerPostModal
-    [posting]="createdAnswerPost"
-    (onCreate)="showAnswers = true; createdAnswerPost = createEmptyAnswerPost()"
-></jhi-answer-post-create-edit-modal>
