+++ src/main/webapp/app/shared/metis/posting-reactions-bar/answer-post-reactions-bar/answer-post-reactions-bar.component.html
@@ -1,4 +1,4 @@
-<div class="me-0 p-1 row">
+<div class="me-0 row justify-content-end">
     <div class="col-auto p-0" *ngFor="let reactionMetaData of reactionMetaDataMap | keyvalue">
         <button
             class="reaction-button clickable"
@@ -6,7 +6,7 @@
             (click)="updateReaction(reactionMetaData.key)"
             [ngbTooltip]="reactionMetaData.value.reactingUsers | reactingUsersOnPosting | async"
         >
-            <ngx-emoji class="emoji" [emoji]="reactionMetaData.key" [size]="16" [set]="'facebook'"> </ngx-emoji>
+            <ngx-emoji class="emoji" [emoji]="reactionMetaData.key" [size]="16"> </ngx-emoji>
             <span class="emoji-count" [class.reaction-count--reacted]="reactionMetaData.value.hasReacted">{{ reactionMetaData.value.count }}</span>
         </button>
     </div>
