+++ src/main/webapp/app/shared/metis/posting-reactions-bar/post-reactions-bar/post-reactions-bar.component.html
@@ -35,9 +35,8 @@
     <div class="col-auto px-0">
         <fa-icon
             [icon]="farSmile"
-            class="reaction-button text-secondary clickable"
+            class="reaction-button me-0 text-secondary clickable"
             [class.open-selector]="showReactionSelector"
-            [class.emoji--dark-mode]="darkModeEmojiPicker"
             (click)="showReactionSelector = !showReactionSelector"
             cdkOverlayOrigin
             #trigger="cdkOverlayOrigin"
@@ -67,29 +66,4 @@
             [set]="'facebook'"
         ></emoji-mart>
     </ng-template>
-    <!-- start discussion -->
-    <div class="col-auto p-0" *ngIf="sortedAnswerPosts.length === 0">
-        <button id="startDiscussionButton" class="reaction-button clickable" (click)="openPostingCreateEditModal.emit()">
-            <ngx-emoji class="emoji" [emoji]="speechBalloonId" [size]="16" [set]="'facebook'"></ngx-emoji>
-            <span class="emoji-count">{{ 'artemisApp.metis.startDiscussion' | artemisTranslate }}</span>
-        </button>
-    </div>
-    <!-- expand answers -->
-    <div class="col-auto p-0" *ngIf="!showAnswers && sortedAnswerPosts.length > 0">
-        <button id="expandAnswersButton" class="reaction-button clickable" (click)="showAnswersChange.emit(true)">
-            <ngx-emoji class="emoji" [emoji]="speechBalloonId" [size]="16" [set]="'facebook'"></ngx-emoji>
-            <span class="emoji-count">{{
-                sortedAnswerPosts.length === 1
-                    ? ('artemisApp.metis.showSingleAnswer' | artemisTranslate)
-                    : ('artemisApp.metis.showMultipleAnswers' | artemisTranslate: { number: sortedAnswerPosts.length })
-            }}</span>
-        </button>
-    </div>
-    <!-- collapse answers -->
-    <div class="col-auto p-0" *ngIf="showAnswers && sortedAnswerPosts.length > 0">
-        <button id="collapseAnswersButton" class="reaction-button clickable" [class.reaction-button--reacted]="showAnswers" (click)="showAnswersChange.emit(false)">
-            <ngx-emoji class="emoji" [emoji]="closeCrossId" [size]="16" [set]="'facebook'"></ngx-emoji>
-            <span class="emoji-count">{{ 'artemisApp.metis.collapseAnswers' | artemisTranslate }}</span>
-        </button>
-    </div>
 </div>
