+++ src/main/webapp/app/shared/metis/posting-reactions-bar/posting-reactions-bar.component.ts
@@ -11,10 +11,6 @@ const PIN_EMOJI_ID = 'pushpin';
 const PIN_EMOJI_UNICODE = '1F4CC';
 const ARCHIVE_EMOJI_ID = 'open_file_folder';
 const ARCHIVE_EMOJI_UNICODE = '1F4C2';
-const SPEECH_BALOON_ID = 'speech_balloon';
-const SPEECH_BALOON_UNICODE = '1F4AC';
-const HEAVY_MULTIPLICATION_ID = 'heavy_multiplication_x';
-const HEAVY_MULTIPLICATION_UNICODE = '2716';
 
 /**
  * event triggered by the emoji mart component, including EmojiData
@@ -45,9 +41,6 @@ interface ReactionMetaDataMap {
 export abstract class PostingsReactionsBarDirective<T extends Posting> implements OnInit, OnChanges, OnDestroy {
     pinEmojiId: string = PIN_EMOJI_ID;
     archiveEmojiId: string = ARCHIVE_EMOJI_ID;
-    speechBalloonId: string = SPEECH_BALOON_ID;
-    closeCrossId: string = HEAVY_MULTIPLICATION_ID;
-
     /*
      * icons (as svg paths) to be used as category preview image in emoji mart selector
      */
@@ -85,14 +78,9 @@ export abstract class PostingsReactionsBarDirective<T extends Posting> implement
      */
     emojisToShowFilter: (emoji: string | EmojiData) => boolean = (emoji) => {
         if (typeof emoji === 'string') {
-            return emoji !== PIN_EMOJI_UNICODE && emoji !== ARCHIVE_EMOJI_UNICODE && emoji !== SPEECH_BALOON_UNICODE && emoji !== HEAVY_MULTIPLICATION_UNICODE;
+            return emoji !== PIN_EMOJI_UNICODE && emoji !== ARCHIVE_EMOJI_UNICODE;
         } else {
-            return (
-                emoji.unified !== PIN_EMOJI_UNICODE &&
-                emoji.unified !== ARCHIVE_EMOJI_UNICODE &&
-                emoji.unified !== SPEECH_BALOON_UNICODE &&
-                emoji.unified !== HEAVY_MULTIPLICATION_UNICODE
-            );
+            return emoji.unified !== PIN_EMOJI_UNICODE && emoji.unified !== ARCHIVE_EMOJI_UNICODE;
         }
     };
